More work
Doesn't work on inventories with inventory sizes less than 9. This is due to the implementation in ContainerChest. I've been avoiding making a new container, want to try to use as much vanilla code as possible.

Let's see if we can't find a way to inject a new container without using proxies! Just need to override the constructor of ContainerChest, that's it.